requirements.txt
httpx>=0.24.0
transformers>=4.36.0
llama-cpp-python>=0.2.11
huggingface-hub>=0.19.0
Pillow>=10.0.0
ollama>=0.1.5
tqdm>=4.66.0
pandas>=2.1.0
torch>=2.1.0
json-repair>=0.8.0
mlx>=0.0.8
mlx-vlm>=0.0.4